chore(channel): Move channel part of tls service to channel module
