+++ tonic/src/transport/channel/endpoint.rs
@@ -1,9 +1,9 @@
 use super::super::service;
+#[cfg(feature = "tls")]
+use super::service::TlsConnector;
 use super::Channel;
 #[cfg(feature = "tls")]
 use super::ClientTlsConfig;
-#[cfg(feature = "tls")]
-use crate::transport::service::TlsConnector;
 use crate::transport::{service::SharedExec, Error, Executor};
 use bytes::Bytes;
 use http::{uri::Uri, HeaderValue};
